refactor: move menus to separate modules
- I created a new `src/menus` directory to house all menu-related modules, organized by functionality.
- Smplified `src/app.py` to serve as the main entry point, importing and calling the `main_menu` from the new structure.
- updated the unit tests to align with the new file organization.
